feat: consistently use predefined network blocks
Previously, we did this for docker swarm but not docker local.

This patch also exposes the chosen subnet to the test via the TestSubnet runtime
environment variable. That way, the test can easily figure out which network
interface is the "data" network and what IP addresses it should use (when asking
the sidecar to assign an IP address).