BUG: null pointer accessing pool.
The k8s runner is initialized during the first Run().

Until there has been at least a single run, c.pool is a null pointer.
I hadn't seen this before, but this causes an exception to be raised if a terminate command is run
prior to running any jobs.

With this patch, a new pool is created, to avoid the problem of having an unitialized runner.

License: MIT
Signed-off-by: Cory Schwartz <cory@protocol.ai>