the big refactor: shipping source, manifests, TESTGROUND_HOME, S… (#849)
## Shipping the test plan source to the daemon

The daemon no longer knows about specific test plans or SDKs, and no longer discovers manifests. The client now submits the source of the test plan to the daemon. We do this through a `multipart/related` request, with three parts:

 * Part 1 (Content-Type: application/json): composition json.
 * Part 2 (Content-Type: application/zip): test plan source.
 * Part 3 (optional, Content-Type: application/zip): linked sdk, see below.

This allows the daemon to be completely agnostic of test plans, and for test plans to be hosted anywhere, not compulsorily on this repo.

## TESTGROUND_HOME

All activity of testground, both client and daemon, revolves around the `TESTGROUND_HOME` directory. It can be specified through the corresponding env variable (`$TESTGROUND_HOME`), or we automatically fall back to `$HOME/testground`. We create the directory and its inner layout automatically, if it doesn't exist.

The inner layout of this directory is as follows:

* `$TESTGROUND_HOME/plans` => where plans are to be hosted. One directory per plan. They can be git working copies, symlinks to local directories, etc. In commands, we reference the plan by its directory name.
* `$TESTGROUND_HOME/sdks` => where SDKs are hosted. This is mostly used by testground developers, as described below, to test changes to an SDK.
* `$TESTGROUND_HOME/data/outputs` => runners deposit test plan outputs here. Of internal use.
* `$TESTGROUND_HOME/data/work` => work directory, for unpacking files, staging source, etc. Of internal use.

The `EnvConfig` now has a `Dirs()` method that returns a `Directories` struct from which testground components can obtain the full path of the above directories.

## Manifests

Manifests are now only a client-side artefact.

Manifests are purely human/machine readable descriptions so that users know what test cases reside in test plan, and which parameters each accepts. They are test plan metadata, useful for building UIs and powering porcelain.

Manifests must now sit on the root of a test plan. As explained above, a test plan can now live anywhere. The only requirement is that it is accompanied by a `manifest.toml` file.

## Describe and list commands

Since the daemon no longer has knowledge of test plans, these commands are now client-side only. They operate on `$TESTGROUND_HOME/plans`.

The `describe` command now only accepts test plans, and enumerates all test cases. The logic to accept test cases has been dropped, as it was poorly implemented.

## SDK linkage

We no longer do any magic `replace`s. Test plans are compiled against a concrete version of the SDK, as determined in their `go.mod`.

However, testground developers tend to modify the SDK itself. And it's extremely handy to be able to modify the SDK locally, and ship that SDK for linkage/replacement/override into the test plan when the build is performed.

This is achieved via a new option: `--link-sdk=<value>`, where `<value>` can be:

1. an absolute path.
2. a path relative to `$PWD`.
3. a path relative to `$TESTGROUND_HOME/sdks`.

(in that order of precedence).

An example of (3) in action would be:

* Inside `$TESTGROUND_HOME/sdks`, I have a symbolic link `sdk-go` that points to my local working copy of the Go SDK.
* When I build/run a test plan, I can supply `--link-sdk=sdk-go`, and the client will pack the SDK source and ship it to the daemon, in part 3 of the multipart request, as explained above.
* The daemon will unpack that SDK, and apply the relevant replace directives (in Go builders) to override the plan's dependency with the shipped SDK.

## Nomenclature of test plans and test cases

Previously, the format to refer to a test case was: `testplan/testcase`. This PR changes the nomenclature to `path/to/test/plan:testcase`.

Everything before the colon is the plan’s location under `$TESTGROUND_HOME/plans`. In the aforementioned case, we expect to find a plan with a `manifest.toml` exactly at: `$TESTGROUND_HOME/path/to/test/plan`. And we expect the `manifest.toml` to contain a test case named `testcase1`.

This new logic makes it possible to unambiguously refer to test plans from within multi-test-plan repos, without confounding slashes for path segments and as a separator between testplan and testcase.

Rationale: I predict folks will host multiple plans in a single repo, e.g. `libp2p/test-plans` may contain `dht`, `pubsub`, etc. test plans.

With #750, we'll add that repo with `testground plan add libp2p git:github.com/libp2p/test-plans.git`, and we'll import multiple test plans at once.