docker:go: rolling go build caches.
This commit introduces rolling/carry-over go build caches as a default
behaviour in docker:go. Build caches images are maintained per test plan +
go version tuple. Every time a build is started, we check if we have an
image from a previous build, and if so, reuse it (clearing /plan,
/sdk, /testground_dep_list first).

If a prior build cache image doesn't exist, we start with a blank go container.

The goal is to carry over the entire /go directory, which contains the
pkg cache. We also set GOCACHE to /go/cache, in order to carry over the
build object cache.

This behaviour can be disabled by setting the `disable_go_build_cache=true`
option in the builder configuration.

The tradeoffs are as follows of using the rolling cache are as follows:

 * good: the rolling cache leads to super-fast go builds, as dependency
   sources and their respective build objects are cached across builds of
   the test plan. For example, if you import the same version of IPFS/
   Lotus/libp2p every time, you will no longer compile from their
   sources, but the go build will use the cached intermediate object
   files. This is akin to running a go build locally.

 * bad: the go toolchain updates last access timestamps on every build
   (https://github.com/golang/go/blob/master/src/cmd/go/internal/cache/cache.go)
   which effectively invalidates Docker layer caching across builds. So
   even if you built exactly the same plan source, no Docker layer
   caches would be hit, because of the timestamps sliding. However, in
   practice, you will be iterating on the test plan source, and under
   that assumption, we have found that the end-to-end build is faster.