feat: rename & fix test plans
This one might needs some explanation: the go tool
assumes that every sub-directory is part of the project,
except the folders that contain a `go.mod` file or have a
name that starts with `_` (or `.` or `testdata`).

In the case of the integrations tests, we DON'T want to
use a `go.mod` because the point of the test is to
verify that GOMOD rewrites are working.

This is why the "best" solution here is to prefix integration with
a `_`.