feat: port example and local exec tests to go tests and github CI (#1537)
Contributes to #1368

Currently, the tests are written in shell, which is hard to maintain, improve, and reuse.
With this PR, we move (some tests) to golang. We'll split this work into a few PR to make it
easier to review.

We call the CLI object from the go tests and start the daemon by creating an actual server object.
We don't go through a testground binary. Which means we can get code coverage and a simpler debugging experience.

- [x] Create an integration test package that relies on golang tool, instead of shell scripts
    - [x] Implement the tooling required: a few run commands, a few `requireXXX` assertions
- [x] port local exec tests
- [x] port docker examples
- [x] make it more practical to use the CLI from a test
    - [x] refactor the generic parser for response to make decoding/printing explicit
    - [x] update the CLI tooling to rely on the passed stdout/stderr rather than printing to the console directly
- [x] Add a basic CI workflow

Co-authored-by: Piotr Galar <piotr.galar@gmail.com>